HistoryActivity : done setting show data transaksi

+++ app/src/main/java/com/example/cafecornerapp/Domain/CartCustomModel.kt
@@ -6,8 +6,8 @@ data class CartCustomModel(
     val cartId: String,
     val productId: String,
     val nama: String,
-    val harga: Int,
+    val harga: Long,
     val kategori: String,
-    val jumlah: Int,
+    val jumlah: Long,
     val imgUrl: String
 ) : Serializable
